Cadastro Eventos - dre e ue todas mandar sem valor no cadastrar, ajuste regra para filtrar tipos de eventos
